+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Linux native crash support ([#734](https://github.com/getsentry/sentry-unity/pull/734))
+- Collect context information synchronously during init to capture it for very early events ([#744](https://github.com/getsentry/sentry-unity/pull/744))
 
 ## 0.16.0